requirements.txt
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jwt
djangorestframework-simplejwt
Flask-JWT-Extended
Pillow==9.3.0
MarkupSafe==2.0.1
django-rest-swagger==2.2.0
django-cors-headers==3.13.0